+++ src/main/java/csc1035/project2/ModuleHandler.java
@@ -56,6 +56,22 @@ public class ModuleHandler {
     }
 
     /**
+     * Method to add staff to a module.
+     *
+     * @param s Staff to teach to a module.
+     * @param m Module for staff to teach.
+     */
+    public void addStaffToModule ( Staff s, Modules m ) {
+        Teach t = new Teach( s.getId(), m.getId() );
+        IController ic = new Controller();
+
+        s.getTeach().add( t );
+        t.setStaff( s );
+        ic.update( t );
+        refreshModuleHandler();
+    }
+
+    /**
      * <code>modules</code> getter method.
      *
      * @return All modules in database.